pyproject.toml
[tool.poetry]
name = "imagesegmentation"
version = "0.1.0"
description = ""
authors = ["Artem Verbov <artem.verbov.osv@fedex.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
clearml = "^1.12.2"
torchvision = "^0.16.2"
opencv-python = "^4.8.1.78"
albumentations = "^1.3.1"
pydantic = "^2.5.3"
lightning = "^2.1.3"
segmentation-models-pytorch = "^0.3.3"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
tensorboard = "^2.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
